+++ modules/eks/ray-cluster/settings.py
@@ -1,7 +1,7 @@
 """Defines the stack settings."""
 
 from abc import ABC
-from typing import Dict, Optional
+from typing import Dict, List, Optional
 
 from pydantic import Field, computed_field
 from pydantic_settings import BaseSettings, SettingsConfigDict
@@ -43,7 +43,7 @@ class SeedFarmerParameters(CdkBaseSettings):
     eks_cluster_admin_role_arn: str
     eks_oidc_arn: str
     service_account_name: str
-    pvc_name: str
+    pvc_name: Optional[str]
     ray_version: str = Field(default="2.30.0")
     ray_cluster_helm_chart_version: str = Field(default="1.1.1")
     image_uri: str = Field(default="rayproject/ray-ml:2.30.0")
@@ -54,7 +54,7 @@ class SeedFarmerParameters(CdkBaseSettings):
     worker_min_replicas: int = Field(default=1)
     worker_max_replicas: int = Field(default=10)
     worker_resources: Dict[str, Dict[str, str]] = Field(default=DEFAULT_POD_RESOURCES)
-    worker_tolerations: Optional[Dict[str, Dict[str, str]]] = Field(default=None)
+    worker_tolerations: List[Dict[str, str]] = Field(default=[])
     tags: Optional[Dict[str, str]] = Field(default=None)