+++ CHANGELOG.md
@@ -11,12 +11,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - adds workflow specific to changes for `requirements-dev.txt` so all static checks are run
 - add `ray-cluster` module based on `kuberay-helm` charts
+- added FSx for Lustre to `ray-on-eks` manifest & persistent volume claim to `ray-cluster` module
+- added worker tolerations to `ray-cluster` module
 
 ### **Changed**
 
 - add integration tests for `sagemaker-studio`
 - bump ecr module version to 1.10.0 to consume auto-delete images feature
-- refactored `ray-on-eks` module to use latest EKS IDF release
 - add service account to kuberay
 - updated `get-modules` workflow to only run tests against changed files in `modules/**`
 - Updated the `sagemaker-templates-service-catalog` module documentation to match the code layout.
@@ -27,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add an example custom `ray-image`
 - document available manifests in readme
 - add permission for SM studio to describe apps when domain resource isolation is enabled
+- updated `ray-on-eks` manifest to use latest EKS IDF release
 
 ## v1.3.0